geo-tracking synchronization

+++ src/UI/Shared/WB.UI.Shared.Extensions/Views/GeoTrackingRecord.cs
@@ -19,7 +19,7 @@ public class GeoTrackingPoint : IPlainStorageEntity<int?>
 {
     [PrimaryKey, Unique, AutoIncrement]
     public virtual int? Id { get; }
-    public double GeoTrackingRecordId { get; set; }
+    public int GeoTrackingRecordId { get; set; }
     public double Latitude { get; set; }
     public double Longitude { get; set; }
     public DateTimeOffset Time { get; set; }
